pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "packing"
readme = "README.md"
license = { file = "LICENSE" }
version = "0.0.1"
requires-python = ">=3.10,<3.13"
dependencies = [
    "absl-py",
    "accelerate",
    "chex",
    "datasets",
    "elkai",
    "huggingface-hub==0.27.0",
    "hydra-core",
    "jax",
    "jumanji",
    "matplotlib",
    "numba",
    "numpy",
    "omegaconf",
    "peft",
    "psutil",
    "requests",
    "scipy",
    "torch",
    "tqdm",
    "transformers",
    "trl",
    "wandb",
]

[tool.hatch.build.targets.wheel]
packages = ["src/packing"]

[tool.isort]
known_third_party = ["wandb"]
